telegram_bot_service/requirements.txt
# python-telegram-bot==20.7.0
# httpx==0.25.2  # Соответствует требованиям python-telegram-bot
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
# requests
fastapi
uvicorn
aiogram